Split `useTodoGroups` mutations out
